meiduo-dev/meiduo_mall/requirements.txt
#iamqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Authlib==1.1.0
billiard==3.6.4.0
bleach==5.0.1
bottle==0.12.23
cached-property==1.5.2
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
cryptography==38.0.1
Deprecated==1.2.13
Django==3.2.15
django-appconf==1.0.5
django-ckeditor==6.3.2
django-compressor==4.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-haystack==3.1.1
django-ipware==4.0.2
django-js-asset==2.0.0
django-mdeditor==0.1.20
django-redis==5.2.0
django-restframework==0.0.1
django-uuslug==2.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
drf-extensions==0.7.1
drf-haystack==1.8.11
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
email-validator==1.3.0
gevent==21.12.0
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.0.1
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.2
jsonpickle==2.2.0
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
mutagen==1.45.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.13.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-alipay-sdk==3.1.0
python-dateutil==2.8.2
python-logstash==0.4.8
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==6.0
QQLoginTool==0.3.0
raven==6.10.0
rcssmin==1.1.0
redis==4.3.4
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.3.0
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
WeRoBot==1.13.1
Whoosh==2.7.4
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
